+++ boot.go
@@ -11,9 +11,9 @@ import (
 	"github.com/deis/builder/pkg/cleaner"
 	"github.com/deis/builder/pkg/conf"
 	"github.com/deis/builder/pkg/gitreceive"
-	"github.com/deis/builder/pkg/gitreceive/storage"
 	"github.com/deis/builder/pkg/healthsrv"
 	"github.com/deis/builder/pkg/sshd"
+	"github.com/deis/builder/pkg/storage"
 	"github.com/deis/builder/pkg/sys"
 	pkglog "github.com/deis/pkg/log"
 	kcl "k8s.io/kubernetes/pkg/client/unversioned"
