+++ pkg/storage/interfaces.go
